Refactor Order component Relates #52
