+++ client/src/Pages/Layout/style.css
@@ -1,5 +1,5 @@
 .loading{
-    margin-top: 50px;
+    margin: 50px auto;
 }
 .page{
     min-height:  calc(100vh - 204px)
